Set property *Norm* to int, instead of double Related to Issue [WTI-279]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@1707 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: a5f861871f3f8efd7827c0326035adc6ef392dd2

+++ Ringtoets/Integration/src/Ringtoets.Integration.Data/Contribution/FailureMechanismContributionItem.cs
@@ -43,7 +43,7 @@ namespace Ringtoets.Integration.Data.Contribution
         /// <summary>
         /// Gets or sets the norm of the complete dike section.
         /// </summary>
-        public double Norm { get; internal set; }
+        public int Norm { get; internal set; }
 
         /// <summary>
         /// Gets the probability space per year for the <see cref="FailureMechanismContribution"/>.
